Fixing Windows build error

+++ src/runtime_src/xdp/profile/writer/native/native_writer.h
@@ -26,9 +26,9 @@ namespace xdp {
   private:
     NativeTraceWriter() = delete ;
 
-    const uint64_t APIBucket = 1 ;
-    const uint64_t readBucket = 2 ;
-    const uint64_t writeBucket = 3 ;
+    const uint32_t APIBucket = 1 ;
+    const uint32_t readBucket = 2 ;
+    const uint32_t writeBucket = 3 ;
 
   protected:
     virtual void writeHeader() ;
